requirements.txt
requests
python-chess
berserk
stockfish
python-dotenv
